+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Added
 
+- `Polyline.Split(List<Vector3> point)`
+- `Polygon.Split(List<Vector3> point)`
+- `Polygon.TrimmedTo(List<Polygon> polygons)`
+- `Vector3.>=`
+- `Vector3.<=`
+- `Plane.Intersects(Plane a, Plane b)`
 - A handful of convenience operators and conversions:
   - implicit `(double X, double Y, double Z)` => `Vector3`
   - implicit `(double X, double Y)` => `Vector3`
@@ -19,6 +25,12 @@
   - `Polygon.UnionAll(params Polygon[] polygons)`
   - `Polygon.Difference(params Polygon[] polygons)`
   - `Polygon.Union(params Polygon[] polygons)`
+
+### Changed
+
+- Some changes to `ContentElement` instance glTF serialization to allow selectability and transformability in the Hypar UI.
+
+### Fixed
 
 ## 0.9.1